Add files via upload
Imagine you're an employee of a pizza company named Contoso Pizza. Your manager has asked you to develop a RESTful pizza inventory management service as a prerequisite for the company's web storefront. The service has to support adding, viewing, modifying, and removing types of pizza—a standardized usage of HTTP action verbs better known as Create, Read, Update, Delete (CRUD).

The .NET CLI is the simplest way to create an ASP.NET Core web API. The CLI is pre-installed when you installed the .NET SDK.

In this unit, you'll use the .NET CLI to create a web API locally. You'll also gain an understanding of the resulting project.